+++ RELEASE_NOTES.md
@@ -1,7 +1,8 @@
 # Next release
 
 - [#420](https://github.com/IAMconsortium/pyam/pull/420) Refactor `data` object to a pandas.Series 
-- [#416](https://github.com/IAMconsortium/pyam/pull/416) Include `meta` in new IamDataFrames returned by aggregation functions
+- [#418](https://github.com/IAMconsortium/pyam/pull/418) Read data from World Bank Open Data Catalogue as IamDataFrame.
+- [#416](https://github.com/IAMconsortium/pyam/pull/416) Include `meta` in new IamDataFrames returned by aggregation functions.
 
 # Release v0.7.0